created endpoint to create sales and salessummary, created dtos for sales and sales summary

+++ src/main/java/com/example/umerautos/dto/SalesSummaryRequestDTO.java
@@ -22,7 +22,7 @@ public class SalesSummaryRequestDTO {
 
     private double totalAmountSummary;
 
-    private  double quantitySoldSummary;
+    private  int quantitySoldSummary;
 
     private List<SaleDTO> saleItems;
